Changes for use in a UR arm with 3.2+ firmware
Added a RobotModeData_V32 for the current most recent data size of 46.

AdditionalInfo is now 8 bytes long rather than 7, don't fail to parse on
this.  I'm probably missing some new byte that isn't in the docs I have
but in the meantime this lets parsing succeed.

A few updates to comments listing interface versions.